More unit tests for simhash

+++ freediscovery/dupdet/simhash.py
@@ -133,7 +133,7 @@ class SimhashDuplicates(BaseEstimator):
         if blocks == 'auto':
             blocks = min(distance*2, 64)
         matches = find_all(self._fit_shash, blocks, distance)
-        matches = np.array(matches)
+        matches = np.array(matches, dtype='uint64')
         return self._fit_shash, cluster_id, matches
